update 16 Dec

+++ app/src/main/java/com/ak/search/app/UpdateReviewAnswer.java
@@ -10,5 +10,5 @@ import java.util.HashMap;
 
 public interface UpdateReviewAnswer {
 
-    public void onReviewUpdate(HashMap<Long,Answers> answers);
+    public void onReviewUpdate(HashMap<Integer,Answers> answers);
 }
